# solutions/Main_Server/swagger/swaggerDoc.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "IUM-TWEB Project 2023-2024 Query Documentation",
    "description": "This is the API documentation for the TWEB project for the years 2023-2024. The project is a football site that retrieves data from two servers, one running Express and the other running SpringBoot. Users can make queries to either of these servers or to the server running on port 3000, which contains all the routes for calls to the two servers."
  },
  "paths": {
    "/getGameEventsByGameId": {
      "get": {
        "summary": "Get game events by game ID",
        "description": "Returns a list of game events for a given game_id. This endpoint is available only on the specified servers.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "Server Express - Main"
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "game_id",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the game for which to get events"
          }
        ],
        "responses": {
          "200": {
            "description": "List of game events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "player_assist_id": {
                        "type": "integer",
                        "nullable": true
                      },
                      "_id": {
                        "type": "string"
                      },
                      "game_event_id": {
                        "type": "string"
                      },
                      "date": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "game_id": {
                        "type": "integer"
                      },
                      "minute": {
                        "type": "integer"
                      },
                      "type": {
                        "type": "string"
                      },
                      "club_id": {
                        "type": "integer"
                      },
                      "player_id": {
                        "type": "integer"
                      },
                      "description": {
                        "type": "string"
                      },
                      "player_in_id": {
                        "type": "integer"
                      },
                      "nome_player_id": {
                        "type": "string"
                      },
                      "nome_player_in_id": {
                        "type": "string"
                      },
                      "nome_player_assist_id": {
                        "type": "string",
                        "nullable": true
                      }
                    }, "example": {
                      "player_assist_id": null,
                      "_id": "65e99ede58797868f77e387b",
                      "game_event_id": "7dca592c3e9ef6d907ecd6f90a1ff1ae",
                      "date": "2012-08-25T00:00:00.000Z",
                      "game_id": 2222597,
                      "minute": 20,
                      "type": "Substitutions",
                      "club_id": 232,
                      "player_id": 45593,
                      "description": ", Injury",
                      "player_in_id": 120630,
                      "nome_player_id": "Welliton",
                      "nome_player_in_id": "Emmanuel Emenike",
                      "nome_player_assist_id": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid game ID supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error description"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Game not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error description"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error retrieving game events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error description"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Single Match Page"
        ]
      }
    },
    "/getGameById": {
      "get": {
        "summary": "Get game by ID",
        "description": "Retrieve details of a game by its ID.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "Server Express - Main"
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "game_id",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the game to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Game details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "description": "Unique identifier for the game"
                    },
                    "game_id": {
                      "type": "integer",
                      "description": "Game ID"
                    },
                    "competition_id": {
                      "type": "string",
                      "description": "Competition ID"
                    },
                    "season": {
                      "type": "integer",
                      "description": "Season year"
                    },
                    "round": {
                      "type": "string",
                      "description": "Round of the competition"
                    },
                    "date": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Date of the game"
                    },
                    "home_club_id": {
                      "type": "integer",
                      "description": "ID of the home club"
                    },
                    "away_club_id": {
                      "type": "integer",
                      "description": "ID of the away club"
                    },
                    "home_club_goals": {
                      "type": "integer",
                      "description": "Goals scored by the home club"
                    },
                    "away_club_goals": {
                      "type": "integer",
                      "description": "Goals scored by the away club"
                    },
                    "home_club_position": {
                      "type": "integer",
                      "description": "Position of the home club in the league"
                    },
                    "away_club_position": {
                      "type": "integer",
                      "description": "Position of the away club in the league"
                    },
                    "home_club_manager_name": {
                      "type": "string",
                      "description": "Name of the home club manager"
                    },
                    "away_club_manager_name": {
                      "type": "string",
                      "description": "Name of the away club manager"
                    },
                    "stadium": {
                      "type": "string",
                      "description": "Stadium where the game was played"
                    },
                    "attendance": {
                      "type": "integer",
                      "description": "Number of attendees"
                    },
                    "referee": {
                      "type": "string",
                      "description": "Name of the referee"
                    },
                    "url": {
                      "type": "string",
                      "description": "URL to the game details"
                    },
                    "home_club_name": {
                      "type": "string",
                      "description": "Name of the home club"
                    },
                    "away_club_name": {
                      "type": "string",
                      "description": "Name of the away club"
                    },
                    "aggregate": {
                      "type": "string",
                      "description": "Aggregate score"
                    },
                    "competition_type": {
                      "type": "string",
                      "description": "Type of competition"
                    }
                  },
                  "example": {
                    "_id": "65e99f1e58797868f78a32bc",
                    "game_id": 2222597,
                    "competition_id": "RU1",
                    "season": 2012,
                    "round": "6. Matchday",
                    "date": "2012-08-25T00:00:00.000Z",
                    "home_club_id": 3725,
                    "away_club_id": 232,
                    "home_club_goals": 2,
                    "away_club_goals": 1,
                    "home_club_position": 2,
                    "away_club_position": 5,
                    "home_club_manager_name": "Stanislav Cherchesov",
                    "away_club_manager_name": "Unai Emery",
                    "stadium": "Akhmat-Arena",
                    "attendance": 21700,
                    "referee": "Vladislav Bezborodov",
                    "url": "https://www.transfermarkt.co.uk/terek-grozny_spartak-moscow/index/spielbericht/2222597",
                    "home_club_name": "RFK Akhmat Grozny",
                    "away_club_name": "FK Spartak Moskva",
                    "aggregate": "2:1",
                    "competition_type": "domestic_league"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid game ID supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error description"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Game not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error description"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Single Match Page"
        ]
      }
    },
    "/getLineUps": {
      "get": {
        "summary": "Retrieve lineups by game ID and club ID",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "Server Express - Main"
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "query",
            "required": false,
            "description": "ID of the game",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "club_id",
            "in": "query",
            "required": false,
            "description": "ID of the club",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "lineups": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": { "type": "string" },
                          "game_lineups_id": { "type": "string" },
                          "game_id": { "type": "integer" },
                          "club_id": { "type": "integer" },
                          "type": { "type": "string" },
                          "number": { "type": "integer" },
                          "player_id": { "type": "integer" },
                          "player_name": { "type": "string" },
                          "team_captain": { "type": "integer" },
                          "position": { "type": "string" }
                        }
                      }
                    }
                  },
                  "example": {
                    "lineups": [
                      {
                        "_id": "65e99f0d58797868f78863d7",
                        "game_lineups_id": "c1301b3ae0c719588ae8732e193b9488",
                        "game_id": 4087929,
                        "club_id": 350,
                        "type": "starting_lineup",
                        "number": 18,
                        "player_id": 61697,
                        "player_name": "Wes Foderingham",
                        "team_captain": 0,
                        "position": "Goalkeeper"
                      },
                      {
                        "_id": "65e99f0d58797868f78863ce",
                        "game_lineups_id": "e244ca9cfe663d38478df6be09f8d5d7",
                        "game_id": 4087929,
                        "club_id": 350,
                        "type": "starting_lineup",
                        "number": 19,
                        "player_id": 128909,
                        "player_name": "Jack Robinson",
                        "team_captain": 0,
                        "position": "Centre-Back"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request, missing required parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No lineups found for the given game_id and club_id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error or no response from the server",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        },
        "tags": ["Single Match Page"]
      }
    },
...
    "/queryPartite": {
      "get": {
        "summary": "Get matches by query parameters",
        "description": "Returns a list of matches based on the provided query parameters. This endpoint is available only on the specified server.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "Server Express - Main"
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            },
            "description": "The season for which to get matches"
          },
          {
            "in": "query",
            "name": "competition_type",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            },
            "description": "The type of competition"
          },
          {
            "in": "query",
            "name": "competition_id",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            },
            "description": "The ID of the competition"
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "default": ""
            },
            "description": "The date for which to get matches"
          }
        ],
        "responses": {
          "200": {
            "description": "List of matches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "game_id": {
                        "type": "integer",
                        "description": "Game ID"
                      },
                      "home_club_name": {
                        "type": "string",
                        "description": "Home club name"
                      },
                      "home_club_goals": {
                        "type": "integer",
                        "description": "Goals scored by the home club",
                        "nullable": true
                      },
                      "away_club_goals": {
                        "type": "integer",
                        "description": "Goals scored by the away club",
                        "nullable": true
                      },
                      "away_club_name": {
                        "type": "string",
                        "description": "Away club name"
                      },
                      "home_club_id": {
                        "type": "integer",
                        "description": "Home club ID"
                      },
                      "away_club_id": {
                        "type": "integer",
                        "description": "Away club ID"
                      }
                    }
                  },
                  "example": [
                    {
                      "game_id": 3844736,
                      "home_club_name": "Associazione Calcio Fiorentina",
                      "home_club_goals": 3,
                      "away_club_goals": 2,
                      "away_club_name": "US Cremonese",
                      "home_club_id": 430,
                      "away_club_id": 2239
                    },
                    {
                      "game_id": 3844743,
                      "home_club_name": "U.S. Salernitana 1919 S.r.l.",
                      "home_club_goals": null,
                      "away_club_goals": 1,
                      "away_club_name": "Associazione Sportiva Roma",
                      "home_club_id": 380,
                      "away_club_id": 12
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error description"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Matches not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error description"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error description"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
...
    "/selectSeasons": {
      "get": {
        "summary": "Get distinct seasons",
        "description": "Returns a list of distinct seasons. This endpoint is available only on the specified server.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "Server Express - Main"
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of distinct seasons",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "description": "Distinct season"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error description"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Seasons not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error description"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error description"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
...
  }
}